Fix build issues after PR #26 rebase onto main
- ImpressSpotlight: bump platform to macOS 26 / iOS 26 to match ImpressLogging
- FilterInput, GlobalSearchPaletteView, NLSearchOverlayView: fix 'some GridRow<some View>' — opaque types in generic params not allowed in Swift 5; use @ViewBuilder + some View
- NLSearchService: reorder LanguageModelSession args (tools before instructions)
- NLSearchFormView: cast feedMaxResults to Int16 (not Int64) for updateSmartSearch
- NLSearchOverlayView: replace .frame(idealHeight:) with .frame(width:) — idealHeight not a valid standalone parameter
- impress-core Item: add `modified` field with serde default; wire through shaped_queries, registry, sqlite_store, store-ffi
- imbib-core: regenerated Swift bindings after filter.rs changes

+++ apps/imbib/PublicationManagerCore/Sources/PublicationManagerCore/SharedViews/NLSearchFormView.swift
@@ -530,7 +530,7 @@ public struct NLSearchFormView: View {
                 return
             }
             let name = feedName.isEmpty ? "AI: \(text.prefix(40))" : feedName
-            RustStoreAdapter.shared.updateSmartSearch(feedID, name: name, query: query, maxResults: Int64(feedMaxResults))
+            RustStoreAdapter.shared.updateSmartSearch(feedID, name: name, query: query, maxResults: Int16(feedMaxResults))
             RustStoreAdapter.shared.updateIntField(id: feedID, field: "refresh_interval_seconds", value: Int64(refreshPreset.rawValue))
             NotificationCenter.default.post(name: .explorationLibraryDidChange, object: nil)
         }
